+++ trailblazer/store/core.py
@@ -1,0 +1,10 @@
+from trailblazer.store.crud.update import UpdateHandler
+
+
+class CoreHandler(
+    UpdateHandler,
+):
+    """Aggregating class for the store API handlers."""
+
+    def __init__(self):
+        pass
